chore: fix copyrights
Add missing copyright headers and update copyright years in files that
have been modified more recently than their copyright states.

note: since updating the copyright is itself a modification, the update
will always be to the current year. E.g. if a file was last modified in
2023, but its copyright is up to 2022; it will be update to the current
year (2024), rather than the year of the previous modification.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>