fix: only one eat.Profile in UnsignedCorim
Bring our implementation in line with the recent [corim spec][1] by
replacing Profiles []eat.Profile field inside UnsignedCorim with a
single profile (multiple profiles have not been allowed since revision
00).

This change required updating test cases that relied on hard-coded CBOR
binary dumps. To make this easier in the future, allow generating them
from YAML descriptions. The test cases are now kept inside external CBOR
files rather than hard-coded in the source (Go's embed mechanism is used
to allow test code to remain the same), along with YAML sources. A
script is provided to re-generate the CBOR from YAML. In the future,
when the structures change, YAML source can be edited to reflect the
change, and the script then used to re-generate the CBOR.

The script relies on an external tool (which the script will install if
its not already present):

	https://github.com/veraison/gen-testcases

note: YAML is used rather than JSON (which we already use for cocli
"templates") because it provides native support for binary
strings/arrays. JSON has no mechanism to distinguish between a string
that is a string, and a string that is a base64 encoding of binary data.
For cocli, the type of the struct field being unmarshaled is used to
make the distinction. For these test cases we do not want to rely on
pre-defined structs, and so conversion is done using only types native
to the formats, hence YAML's support for byte data via !!binary tag is
essential.

[1]: https://datatracker.ietf.org/doc/html/draft-ietf-rats-corim-04

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>